+++ Checkout/Entities/Offer.cs
@@ -2,7 +2,7 @@ namespace Checkout.Entities;
 
 public class Offer
 {
-    public Offer(string sku, int quantity, decimal discountPrice )
+    public Offer(string sku, int quantity, double discountPrice )
     {
         this.Sku = sku;
         this.Quantity = quantity;
@@ -11,5 +11,5 @@ public class Offer
 
     public string Sku { get; set; }
     public int Quantity { get; set; }
-    public decimal DiscountPrice { get; set; }
+    public double DiscountPrice { get; set; }
 }
